fix: cannot delete the same team

+++ src/Core/Application/Company/Teams/DeleteTeamRequest.cs
@@ -1,3 +1,4 @@
+using Ticket360.Domain.Common.Events;
 using Ticket360.Domain.Company;
 
 namespace Ticket360.Application.Company.Teams;
@@ -10,20 +11,25 @@ public class DeleteTeamRequest : IRequest<Guid>
 
 public class DeleteTeamRequestHandler : IRequestHandler<DeleteTeamRequest, Guid>
 {
-    private readonly IRepositoryWithEvents<Team> _repository;
+    private readonly IRepository<Team> _repository;
     private readonly IStringLocalizer _t;
 
-    public DeleteTeamRequestHandler(IRepositoryWithEvents<Team> repository, IStringLocalizer<DeleteTeamRequestHandler> localizer) =>
+    public DeleteTeamRequestHandler(IRepository<Team> repository, IStringLocalizer<DeleteTeamRequestHandler> localizer) =>
         (_repository, _t) = (repository, localizer);
 
     public async Task<Guid> Handle(DeleteTeamRequest request, CancellationToken cancellationToken)
     {
         var team = await _repository.GetByIdAsync(request.Id, cancellationToken);
 
-        _ = team ?? throw new NotFoundException(_t["Team {0} Not Found."]);
+        if (team.DeletedOn != null)
+        {
+            var errorMessage = _t["Team {0} has already been deleted."];
+            throw new NotFoundException(string.Format(errorMessage, request.Id));
+        }
 
+        team.DomainEvents.Add(EntityDeletedEvent.WithEntity(team));
         await _repository.DeleteAsync(team, cancellationToken);
-
         return request.Id;
+
     }
 }
